+++ sdk/crypto/basic/crypto_keyexchange.hpp
@@ -122,7 +122,7 @@ class crypto_keyexchange {
      *
      *          inconsistent concatenation
      *          - tls_group_secp256r1mlkem768  secp256r1 || mlkem768
-     *          - tls_group_x25519mlkem768     mlkem768 || x25519 (due to historical reasons)
+     *          - tls_group_x25519mlkem768     mlkem768  || x25519 (due to historical reasons)
      *          - tls_group_secp384r1mlkem1024 secp384r1 || mlkem1024
      */
     return_t keygen(tls_group_t group, crypto_key* key, const char* kid);