+++ test/tls/test_construct_tls.cpp
@@ -948,9 +948,9 @@ void test_construct_tls13_mlkem() {
         test_construct_tls_routine(item, "MLKEM768");
         test_construct_tls_routine(item, "MLKEM1024");
 
-        // test_construct_tls_routine(item, "SecP256r1MLKEM768");
+        test_construct_tls_routine(item, "SecP256r1MLKEM768");
         test_construct_tls_routine(item, "X25519MLKEM768");
-        // test_construct_tls_routine(item, "SecP384r1MLKEM1024");
+        test_construct_tls_routine(item, "SecP384r1MLKEM1024");
     }
     tlsadvisor->set_default_tls_groups();
 #else